+++ scipy/integrate/meson.build
@@ -140,6 +140,7 @@ subdir('tests')
 py3.install_sources([
     '__init__.py',
     '_bvp.py',
+    '_lebedev.py',
     '_ode.py',
     '_quad_vec.py',
     '_quadrature.py',
