+++ scipy/special/__init__.py
@@ -548,7 +548,6 @@ orthogonal polynomials:
    roots_sh_chebyt   -- Gauss-Chebyshev (first kind, shifted) quadrature.
    roots_sh_chebyu   -- Gauss-Chebyshev (second kind, shifted) quadrature.
    roots_sh_jacobi   -- Gauss-Jacobi (shifted) quadrature.
-   roots_lebedev     -- Lebedev quadrature.
 
 The functions below, in turn, return the polynomial coefficients in
 ``orthopoly1d`` objects, which function similarly as `numpy.poly1d`.
@@ -751,7 +750,6 @@ Other special functions
    spence      -- Spence's function, also known as the dilogarithm.
    zeta        -- Riemann zeta function.
    zetac       -- Riemann zeta function minus 1.
-   softplus    -- Softplus function.
 
 Convenience functions
 ---------------------
@@ -843,7 +841,6 @@ from ._ellip_harm import (
     ellip_normal
 )
 from ._lambertw import lambertw
-from ._lebedev import roots_lebedev
 from ._spherical_bessel import (
     spherical_jn,
     spherical_yn,
